requirements.txt
asgiref==3.2.3
bootstrap4==0.1.0
Django==3.0.3
django-crispy-forms==1.8.1
gunicorn==20.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
